created routes files and cleaned code, still works
